use gsize instead of int where appropriate (64-bit cleanliness fix)
Sun Sep 29 12:15:44 2002  Manish Singh  <yosh@gimp.org>

        * tests/mainloop-test.c: use gsize instead of int where appropriate
        (64-bit cleanliness fix)

+++ tests/mainloop-test.c
@@ -54,10 +54,10 @@ struct _TestData
 static void cleanup_crawlers (GMainContext *context);
 
 gboolean
-read_all (GIOChannel *channel, char *buf, int len)
+read_all (GIOChannel *channel, char *buf, gsize len)
 {
-  int bytes_read = 0;
-  int count;
+  gsize bytes_read = 0;
+  gsize count;
   GIOError err;
 
   while (bytes_read < len)
@@ -78,10 +78,10 @@ read_all (GIOChannel *channel, char *buf, int len)
 }
 
 gboolean
-write_all (GIOChannel *channel, char *buf, int len)
+write_all (GIOChannel *channel, char *buf, gsize len)
 {
-  int bytes_written = 0;
-  int count;
+  gsize bytes_written = 0;
+  gsize count;
   GIOError err;
 
   while (bytes_written < len)
